Remade dialogue system and made other small changes
The dialogue system is significantly more complex. The player can choose a topic and then respod to what the npc says and the dialogue can continue indefinately. The enums for races and relationships have been changed from numbers to strings. The print to journal function will wrap words to the next line.